feat(invoices): refactor invoice creation and delivery services
- Introduced new helper functions for invoice creation and delivery in `invoice-creation.helpers.ts` and `invoice.delivery.helpers.ts`.
- Implemented validation logic for invoice creation, ensuring proper client and matter associations.
- Enhanced invoice delivery process with Stripe integration, including error handling and reconciliation for voided invoices.
- Updated invoice service to utilize new helper functions, improving code organization and readability.
- Added pagination schema to OpenAPI validation for better API documentation.
- Implemented webhook handling for Stripe invoice events, including syncing seat counts for upcoming and created invoices.
- Improved error logging and handling throughout the invoice processing workflow.
