+++ src/modules/invoices/handlers.ts
@@ -1,6 +1,6 @@
 import type { routes } from '@/modules/invoices/routes';
+import { serializeInvoice, serializePaginatedInvoices } from '@/modules/invoices/serializers/invoice.serializer';
 import { invoiceService } from '@/modules/invoices/services/invoice.service';
-import { InvoiceCreated } from '@/shared/events/definitions';
 import { invoiceDeliveryService } from '@/modules/invoices/services/invoice.delivery.service';
 import type { AppRouteHandler } from '@/shared/types/hono';
 import { getServiceContext, createServiceContext } from '@/shared/types/service-context';
@@ -8,28 +8,12 @@ import { db } from '@/shared/database';
 
 const createInvoiceHandler: AppRouteHandler<typeof routes.createInvoiceRoute> = async (c) => {
   const { practice_id: organizationId } = c.req.valid('param');
-  const baseCtx = { ...getServiceContext(c), organizationId };
+  const ctx = { ...getServiceContext(c), organizationId };
   const data = c.req.valid('json');
 
-  const result = await db.transaction(async (tx) => {
-    const ctx = createServiceContext(baseCtx, tx);
+  const result = await invoiceService.createInvoice({ data }, ctx);
 
-    const invoiceCreated = await invoiceService.createInvoice({ data }, ctx);
-
-    // If creation succeeded, emit InvoiceCreated within same transaction
-    await ctx.emit(InvoiceCreated, {
-      invoice_id: invoiceCreated.id,
-      organization_id: ctx.organizationId,
-      client_id: invoiceCreated.client_id,
-      matter_id: invoiceCreated.matter_id ?? null,
-      invoice_number: invoiceCreated.invoice_number ?? null,
-      total: invoiceCreated.total,
-    });
-
-    return invoiceCreated;
-  });
-
-  return c.json(result, 201);
+  return c.json(serializeInvoice(result), 201);
 };
 
 const listInvoicesHandler: AppRouteHandler<typeof routes.listInvoicesRoute> = async (c) => {
@@ -39,7 +23,7 @@ const listInvoicesHandler: AppRouteHandler<typeof routes.listInvoicesRoute> = as
 
   const result = await invoiceService.listInvoices({ filters: query }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializePaginatedInvoices(result), 200);
 };
 
 const getInvoiceHandler: AppRouteHandler<typeof routes.getInvoiceRoute> = async (c) => {
@@ -48,20 +32,17 @@ const getInvoiceHandler: AppRouteHandler<typeof routes.getInvoiceRoute> = async 
 
   const result = await invoiceService.getInvoiceById({ id }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializeInvoice(result), 200);
 };
 
 const updateInvoiceHandler: AppRouteHandler<typeof routes.updateInvoiceRoute> = async (c) => {
   const { invoice_id: id, practice_id: organizationId } = c.req.valid('param');
-  const baseCtx = { ...getServiceContext(c), organizationId };
+  const ctx = { ...getServiceContext(c), organizationId };
   const data = c.req.valid('json');
 
-  const result = await db.transaction(async (tx) => {
-    const ctx = createServiceContext(baseCtx, tx);
-    return await invoiceService.updateInvoice({ id, data }, ctx);
-  });
+  const result = await invoiceService.updateInvoice({ id, data }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializeInvoice(result), 200);
 };
 
 const deleteInvoiceHandler: AppRouteHandler<typeof routes.deleteInvoiceRoute> = async (c) => {
@@ -81,7 +62,7 @@ const sendInvoiceHandler: AppRouteHandler<typeof routes.sendInvoiceRoute> = asyn
   const ctx = { ...getServiceContext(c), organizationId };
   const result = await invoiceDeliveryService.sendInvoice({ id }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializeInvoice(result), 200);
 };
 
 const syncInvoiceHandler: AppRouteHandler<typeof routes.syncInvoiceRoute> = async (c) => {
@@ -90,7 +71,7 @@ const syncInvoiceHandler: AppRouteHandler<typeof routes.syncInvoiceRoute> = asyn
 
   const result = await invoiceDeliveryService.syncInvoice({ id }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializeInvoice(result), 200);
 };
 
 const voidInvoiceHandler: AppRouteHandler<typeof routes.voidInvoiceRoute> = async (c) => {
@@ -98,7 +79,7 @@ const voidInvoiceHandler: AppRouteHandler<typeof routes.voidInvoiceRoute> = asyn
   const ctx = { ...getServiceContext(c), organizationId };
   const result = await invoiceDeliveryService.voidInvoice({ id }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializeInvoice(result), 200);
 };
 
 const getClientInvoicesHandler: AppRouteHandler<typeof routes.getClientInvoicesRoute> = async (c) => {
@@ -108,7 +89,7 @@ const getClientInvoicesHandler: AppRouteHandler<typeof routes.getClientInvoicesR
 
   const result = await invoiceService.listClientInvoices({ filters: query }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializePaginatedInvoices(result), 200);
 };
 
 const getClientInvoiceDetailHandler: AppRouteHandler<typeof routes.getClientInvoiceDetailRoute> = async (c) => {
@@ -117,7 +98,7 @@ const getClientInvoiceDetailHandler: AppRouteHandler<typeof routes.getClientInvo
 
   const result = await invoiceService.getClientInvoiceDetail({ invoiceId: id }, ctx);
 
-  return c.json(result, 200);
+  return c.json(serializeInvoice(result), 200);
 };
 
 export const handlers = {
