+++ src/modules/invoices/services/invoice.utils.ts
@@ -5,11 +5,10 @@ import type { db } from '@/shared/database';
 
 export const calculateInvoiceTotals = (lineItems: InvoiceLineItemInput[]): InvoiceTotals => {
   const subtotal = lineItems.reduce((acc, item) => acc + item.quantity * item.unit_price, 0);
-  const tax_amount = 0;
-  const discount_amount = 0;
-  const total = subtotal + tax_amount - discount_amount;
+  // Tax and discounts are intentionally fixed to zero until those rules exist.
+  const total = subtotal;
 
-  return { subtotal, tax_amount, discount_amount, total, amount_due: total };
+  return { subtotal, tax_amount: 0, discount_amount: 0, total, amount_due: total };
 };
 
 export const extractPayoutMeteredFeeCents = (invoiceTxs: SelectBillingTransaction[]): number | null => {
